+++ R/reexports.R
@@ -1,23 +1,4 @@
 #' @export
-#' @importFrom datawizard adjust
-datawizard::adjust
-
-#' @export
-#' @importFrom datawizard data_rescale
-datawizard::data_rescale
-
-#' @export
-#' @importFrom datawizard normalize
-datawizard::normalize
-
-#' @export
-#' @importFrom datawizard ranktransform
-datawizard::ranktransform
-
-#' @export
 #' @importFrom datawizard standardize
 datawizard::standardize
 
-#' @export
-#' @importFrom datawizard unstandardize
-datawizard::unstandardize